refactor(quote): modularizar formulario de cotización y añadir hook useCarQuoteForm
- Se extrae la lógica del formulario al hook personalizado `useCarQuoteForm`.
- Se dividen los campos del formulario en el componente `CarQuoteFormFields`.
- Se actualizan las referencias en `CarForm`, `header` y `footer` para usar la nueva estructura.
- Se añade el componente `PriceTable` para la visualización de precios.
- Se elimina el archivo `carQuoteForm.jsx` por redundancia.
